Introduce flux dispatcher and Howls store
Add a Flux dispatcher with three defined actions, appInit, newHowl, and
newHowlCreated.

Add a Howls store, which uses a Howls collection to store data.

Update New Howls view to dispatch the newHowl action in response to the
submit DOM event.  The Howls store responds to a newHowl action, and
creates a new howl in the store.  Upon success, the Howls store
dispatches a newHowlCreated action.  The New Howls view resets the new
howl input in response to the newHowlCreated action.